+++ src/aux-server/aux-web/aux-player/interaction/PlayerInteractionManager.ts
@@ -1,4 +1,12 @@
-import { Vector3, Intersection, Object3D, OrthographicCamera } from 'three';
+import {
+    Vector3,
+    Intersection,
+    Object3D,
+    OrthographicCamera,
+    Quaternion,
+    Euler,
+    Matrix4,
+} from 'three';
 import { ContextMenuAction } from '../../shared/interaction/ContextMenuEvent';
 import {
     Bot,
@@ -7,19 +15,32 @@ import {
     ON_FOCUS_ENTER_ACTION_NAME,
     ON_ANY_FOCUS_ENTER_ACTION_NAME,
     ON_ANY_FOCUS_EXIT_ACTION_NAME,
+    PortalType,
+    calculateGridScale,
+    BotTags,
+    hasValue,
 } from '@casual-simulation/aux-common';
 import { IOperation } from '../../shared/interaction/IOperation';
 import { BaseInteractionManager } from '../../shared/interaction/BaseInteractionManager';
 import { GameObject } from '../../shared/scene/GameObject';
 import { AuxBot3D } from '../../shared/scene/AuxBot3D';
 import { PlayerBotClickOperation } from './ClickOperation/PlayerBotClickOperation';
-import { Input, ControllerData, InputMethod } from '../../shared/scene/Input';
+import {
+    Input,
+    ControllerData,
+    InputMethod,
+    InputState,
+    MouseButtonId,
+} from '../../shared/scene/Input';
 import { appManager } from '../../shared/AppManager';
 import { Simulation } from '@casual-simulation/aux-vm';
 import { DraggableGroup } from '../../shared/interaction/DraggableGroup';
 import flatMap from 'lodash/flatMap';
 import { InventoryContextGroup3D } from '../scene/InventoryContextGroup3D';
-import { isObjectVisible } from '../../shared/scene/SceneUtils';
+import {
+    isObjectVisible,
+    objectForwardRay,
+} from '../../shared/scene/SceneUtils';
 import { CameraRigControls } from '../../shared/interaction/CameraRigControls';
 import { CameraControls } from '../../shared/interaction/CameraControls';
 import {
@@ -31,6 +52,11 @@ import { PlayerGame } from '../scene/PlayerGame';
 import { DimensionGroup3D } from '../../shared/scene/DimensionGroup3D';
 import { Grid3D } from '../Grid3D';
 import { PlayerPageSimulation3D } from '../scene/PlayerPageSimulation3D';
+import { PlayerSimulation3D } from '../scene/PlayerSimulation3D';
+import { InventorySimulation3D } from '../scene/InventorySimulation3D';
+import { Physics } from '../../shared/scene/Physics';
+import { Simulation3D } from '../../shared/scene/Simulation3D';
+import isEqual from 'lodash/isEqual';
 
 export class PlayerInteractionManager extends BaseInteractionManager {
     // This overrides the base class Game.
@@ -275,6 +301,206 @@ export class PlayerInteractionManager extends BaseInteractionManager {
         }
 
         return [mainCameraRigControls, invCameraRigControls];
+    }
+
+    // This function is kinda the worst but should be fine
+    // as long as performance doesn't become an issue.
+    protected _updatePlayerBotTags() {
+        const input = this._game.getInput();
+        const pagePos = this._game.getInput().getMousePagePos();
+        const draggableGroups = this.getDraggableGroups();
+        const viewports = this._game.getViewports();
+
+        let inputList = [
+            'keyboard',
+            'mousePointer',
+            'touch',
+            ...input.controllers.map(c => `${c.inputSource.handedness}Pointer`),
+        ];
+
+        let inputUpdate = {
+            inputList: inputList,
+        } as BotTags;
+
+        for (let key of input.getKeys()) {
+            checkInput(key.state, `keyboard_${key.key}`, inputUpdate);
+        }
+
+        const leftState = input.getButtonInputState(MouseButtonId.Left);
+        const rightState = input.getButtonInputState(MouseButtonId.Right);
+        const middleState = input.getButtonInputState(MouseButtonId.Middle);
+        checkInput(leftState, 'mousePointer_left', inputUpdate);
+        checkInput(rightState, 'mousePointer_right', inputUpdate);
+        checkInput(middleState, 'mousePointer_middle', inputUpdate);
+
+        for (let i = 0; i < 5; i++) {
+            const touch = input.getTouchData(i);
+            if (touch) {
+                checkInput(touch.state, `touch_${i}`, inputUpdate);
+            } else {
+                inputUpdate[`touch_${i}`] = null;
+            }
+        }
+
+        for (let sim of this._game.getSimulations()) {
+            const rig = sim.getMainCameraRig();
+            const cameraWorld = new Vector3();
+            cameraWorld.setFromMatrixPosition(rig.mainCamera.matrixWorld);
+            const cameraRotation = new Euler();
+            cameraRotation.setFromRotationMatrix(rig.mainCamera.matrixWorld);
+            const [portal, gridScale, inverseScale] = portalInfoForSim(sim);
+
+            let update = {
+                ...inputUpdate,
+                [`${portal}CameraPositionX`]: cameraWorld.x * inverseScale,
+                [`${portal}CameraPositionY`]: -cameraWorld.z * inverseScale,
+                [`${portal}CameraPositionZ`]: cameraWorld.y * inverseScale,
+                [`${portal}CameraRotationX`]: cameraRotation.x,
+                [`${portal}CameraRotationY`]: cameraRotation.z,
+                [`${portal}CameraRotationZ`]: cameraRotation.y,
+            };
+
+            for (let i = 0; i < draggableGroups.length; i++) {
+                const group = draggableGroups[i];
+                const objects = group.objects;
+                const camera = group.camera;
+                const viewport = group.viewport;
+
+                if (sim.getMainCameraRig().viewport !== group.viewport) {
+                    continue;
+                }
+
+                if (
+                    !Input.pagePositionOnViewport(pagePos, viewport, viewports)
+                ) {
+                    // Page position is not on or is being obstructed by other viewports.
+                    // Ignore this draggable group.
+                    continue;
+                }
+
+                const screenPos = Input.screenPositionForViewport(
+                    pagePos,
+                    viewport
+                );
+                const ray = Physics.rayAtScreenPos(screenPos, camera);
+                const mat = new Matrix4();
+                mat.lookAt(
+                    ray.origin,
+                    ray.direction.add(ray.origin),
+                    new Vector3(0, 1, 0)
+                );
+                const worldRotation = new Euler();
+                worldRotation.setFromRotationMatrix(mat);
+
+                Object.assign(update, {
+                    [`mousePointerPositionX`]: ray.origin.x * inverseScale,
+                    [`mousePointerPositionY`]: -ray.origin.z * inverseScale,
+                    [`mousePointerPositionZ`]: ray.origin.y * inverseScale,
+                    [`mousePointerRotationX`]: worldRotation.x,
+                    [`mousePointerRotationY`]: worldRotation.z,
+                    [`mousePointerRotationZ`]: worldRotation.y,
+                    [`mousePointerPortal`]: portal,
+                });
+            }
+
+            if (!(sim instanceof PlayerPageSimulation3D)) {
+                for (let controller of input.controllers) {
+                    const ray = objectForwardRay(controller.ray);
+                    const mat = new Matrix4();
+                    mat.lookAt(
+                        ray.origin,
+                        ray.direction.add(ray.origin),
+                        new Vector3(0, 1, 0)
+                    );
+                    const worldRotation = new Euler();
+                    worldRotation.setFromRotationMatrix(mat);
+                    const hand = controller.inputSource.handedness;
+
+                    let inputStates = {};
+                    checkInput(
+                        controller.primaryInputState,
+                        `${hand}Pointer_primary`,
+                        inputStates
+                    );
+                    checkInput(
+                        controller.squeezeInputState,
+                        `${hand}Pointer_squeeze`,
+                        inputStates
+                    );
+
+                    Object.assign(update, {
+                        ...inputStates,
+                        [`${hand}PointerPositionX`]:
+                            ray.origin.x * inverseScale,
+                        [`${hand}PointerPositionY`]:
+                            -ray.origin.z * inverseScale,
+                        [`${hand}PointerPositionZ`]:
+                            ray.origin.y * inverseScale,
+                        [`${hand}PointerRotationX`]: worldRotation.x,
+                        [`${hand}PointerRotationY`]: worldRotation.z,
+                        [`${hand}PointerRotationZ`]: worldRotation.y,
+                        [`${hand}PointerPortal`]: portal,
+                    });
+                }
+            }
+
+            const userBot = sim.simulation.helper.userBot;
+            for (let key in update) {
+                const userValue = userBot.tags[key];
+                const updateValue = update[key];
+                if (
+                    userValue === updateValue ||
+                    (!hasValue(userValue) && !hasValue(updateValue)) ||
+                    (hasValue(userValue) &&
+                        hasValue(updateValue) &&
+                        isEqual(userValue, updateValue)) ||
+                    (typeof userValue === 'number' &&
+                        typeof updateValue === 'number' &&
+                        isNaN(userValue) &&
+                        isNaN(updateValue))
+                ) {
+                    delete update[key];
+                }
+            }
+
+            if (Object.keys(update).length > 0) {
+                sim.simulation.helper.updateBot(sim.simulation.helper.userBot, {
+                    tags: update,
+                });
+            }
+        }
+
+        function portalInfoForSim(sim: Simulation3D) {
+            let portal: PortalType;
+            let gridScale: number;
+            if (sim instanceof PlayerPageSimulation3D) {
+                portal = 'page';
+                gridScale = sim.pageConfig.gridScale;
+            } else if (sim instanceof InventorySimulation3D) {
+                portal = 'inventory';
+                gridScale = sim.inventoryConfig.gridScale;
+            }
+            let inverseScale = 1 / gridScale;
+
+            return [portal, gridScale, inverseScale] as const;
+        }
+
+        function checkInput(state: InputState, name: string, update: any) {
+            if (!state) {
+                return false;
+            }
+            if (state.isDownOnFrame(input.time.frameCount)) {
+                inputUpdate[name] = 'down';
+                return true;
+            } else if (state.isHeldOnFrame(input.time.frameCount)) {
+                inputUpdate[name] = 'held';
+                return true;
+            } else if (state.isUpOnFrame(input.time.frameCount)) {
+                inputUpdate[name] = null;
+                return true;
+            }
+            return false;
+        }
     }
 
     protected _contextMenuActions(